refactor(store): extract store/profile repository from database.py
Move 9 municipality energy profile data funcs (ElCom tariffs, profiles,
Sonnendach) into store/profile.py, re-exported from database.py so
db.<fn>() keeps working. Continues store/ranking precedent (Candidate #4).
No behavior change; full suite + ruff green.

Closes #75
